distro: introduce "process-containment" feature
Make the softwarecontainer package optional, by adding a DISTRO_FEATURE
flag called "process-containment".  In order to keep compatibility with
the existing images, this feature is enabled by default for the "pelux"
images.

Distributions which disable this feature will not have the
softwarecontainer package installed, and therefore might not require the
meta-virtualization layer, which is being pulled in just to satisfy the
softwarecontainer dependencies.

Fixes #33