swupdate: Overrides the raspberrypi3 configuration for minimal config
This commit configures SWUpdate with a minimal configuration that
doesn't include the Mongoose embedded webserver. The "suricatta" mode is
enabled for future Hawkbit integration.

A couple of partitioning tools are also added as dependencies of
swupdate to enable dynamic repartitioning while updating on the raspberry
pi.

No signature verification is enforced by this configuration. Therefore,
this config shouldn't be used in production as-is.

Signed-off-by: Florent Revest <revestflo@gmail.com>