Adds a core-image-pelux-minimal-update recipe to generate a .swu bundle
This patch creates a core-image-pelux-minimal-update recipe that can be
bitbaked to obtain a .swu file of a core-image-pelux-minimal rootfs for
the Raspberry Pi.

The .swu bundle contains a sw-description file describing the operations
to be executed during the upgrade. It is split in two parts that can be
executed depending on the partition that should be flashed. Each part is
divided in three operations: - A pre-installation lua script
repartitions the SDCard to make sure the partition layout supports an
A/B update scheme.  - A compressed ext3 rootfs is flashed on the
partition of interest.  - An U-Boot envvar is set to specify atomically
the new rootfs to be used.

The *.swu resulting of a bitbake core-image-pelux-minimal-update can be
found in build/tmp/deploy/images/raspberrypi3/ and can be executed on
the target using a command such as:

swupdate -v -H raspberrypi3:1.0 -e stable,alt -f /etc/swupdate.cfg -i
update.swu

Signed-off-by: Florent Revest <revestflo@gmail.com>