Generify SWUpdate artifacts targets
We rely on the meta-swupdate's swupdate.bbclass to generate .swu update
artifacts for a given image. However, this bbclass requires to have one
recipe per artifact. Since we have four different targets, we need to
factorize the work between those four recipes.

This patch introduces a common pelux-image-update.inc file that can be
included by all the artifacts recipes to generate appropriate artifacts
from a set of generic config files.

This inc file finds sw-descriptor and scripts in MACHINE-specific
subdirectories of the pelux-image-update/ directory. In order to be kept
generic, those sw-desciptors can refer to the "PELUX_IMAGE" token that
will be replaced by the name of the necessary PELUX image dynamically.

Signed-off-by: Florent Revest <revestflo@gmail.com>