mopidy: add recipe
Mopidy is a Python media playback service that is used as a reference
implementation for PELUX. It can be extended with various plugins
offering MPD or Spotify integration.

This recipe fetches and compiles Mopidy directly from pypi. It also
provides a couple of default options necessary to let Mopidy work
correctly out of the box (a default media directory and also an
integration with Pulseaudio). Last but not least, it also includes a
systemd service to start Mopidy automatically at boot time.

Signed-off-by: Florent Revest <revestflo@gmail.com>