neptune3: Force to run as single process for rpi
With neptune3, multiple applications in neptune ui
runs as separate processes. On rpi this causes some
performance issues since it doesn't have a very good
GPU. This causes a very low frame rate as well as
issues with correct rendering of the applications.
A workaround is added which runs neptune3 ui as a
single process. This fixes the performance and
rendering issues on rpi. This is a temporary fix,
until a solution is provided in neptune for this
issue.

Signed-off-by: Tariq Ansari <tansari@luxoft.com>