grub-efi: add fstab to meta-intel to mount EFI partition
An fstab file was missing in PELUX to mount EFI partition. This resulted
in swupdate not able to edit grubenv environment block which is located
on EFI boot partition. Previously, a workaround was used to mount the
EFI partition using grubenv-copy.service at boot time. This workaround
was error prone, so it has now been replaced by adding fstab file
for meta-intel layer.

Signed-off-by: Tariq Ansari <tansari@luxoft.com>