neptune: run qtwebengine in single process mode
QtWebEngine crashes trying load some websites, e.g. youtube.com.

This is an issue of the QtWebEngine. Running it outside the sandbox
and in the single process mode is a temporary workaround.

There is a bug for this:
https://bugreports.qt.io/browse/AUTOSUITE-926

--single-process is already passed to the Neptune executable itself for
raspberrypi.See FORCE_SINGLE_PROCESS in
layers/b2qt/recipes-qt/automotive/neptune3-ui_git.bbappend .
However, that is different from adding an enviroment variable to
QTWEBENGINE_CHROMIUM_FLAGS. When running the QTWEBENGINE with
--single-process and withouth sandbox, the youtube app is fixed. Also
this fix is for all variants.

Signed-off-by: Fisnik Hajredini <fhajredini@luxoft.com>