ARP: Add multiconfig files for sos and uos
Add multiconfig files for sos and uos under conf/variant/arp. These files
are needed when building the ServiceOS and User OS versions of PELUX for
ARP.

To use them they need to be copied to the `${TOPDIR}/conf/multiconfig`
directory and `BBMULTICONFIG = "sos uos"` needs to be added to local.conf.

Signed-off-by: Viktor Sjölind <vsjolind@luxoft.com>