core-image-pelux-sos: Add ServiceOS image
Add image that builds a Service OS version of PELUX which boots through
the acrn-hypervisor. The image also installs a PELUX VM which
automatically boots at startup.

In order to run this image multiconfig for SOS and UOS need to be setup.
Look at conf/variant/arp/multiconfig for inspiration.
Documentation for how to set this up can be found at
https://www.yoctoproject.org/docs/latest/mega-manual/mega-manual.html#var-BBMULTICONFIG

When multiconfig is setup, build using the following command:
`bitbake multiconfig:sos:core-image-pelux-sos`

Signed-off-by: Viktor Sjölind <vsjolind@luxoft.com>