Use WAYLAND_DISPLAY env var in neptune & softwarecontainer
Wayland socket file name was hardcode in softwarecontainer which
resulted in failure if the wayland socket file name was changed
on the file system. Neptune provides WAYLAND_DISPLAY env variable
to define the name of the WAYLAND socket. Use this variable to
set the socket name in neptune and softwarecontainer.

Signed-off-by: Tariq Ansari <tansari@luxoft.com>