Use PARTUUID to boot intel and ARP targets
NUC and ARP were assumed to boot from /dev/sda.
However, the boot device could be other than the
usb. In that case hardcoded device /dev/sda would
fail to boot. The solution is to use manually
generated PARTUUIDs hardcoded in wks file instead
of hardcoded device names, which could change. The
PARTUUIDs will remain the same even if the boot
device is changed. This also affected the swupdate
which also assumed the device to be /dev/sda.
Software update is modifed to use PARTLABEL to
detect the boot device. Since PARTUUID and
PARTLABEL have one to one mapping, PARTUUID
can be used in software update scripts. However,
for better readability reasons PARTLABEL are used.

Signed-off-by: Tariq Ansari <tansari@luxoft.com>