local.conf.sample: add smoketest to qemu config
Currently, when one needs to run smoketests on qemu, the
local.conf file of the qemu-x86-64_nogfx variant needs to
be manually modified to enable running of the tests. The
tests and the configuration required to run these tests
are located in conf/test-scripts/local.conf.appendix file
in pelux-manifest. For running nightly smoke tests the
configuration from this file is automatically extracted
and appended to the qemu local.conf, but to run the tests
locally, one needs to copy the config to the local.conf.
This commit adds this configuration to local.conf.sample
of qemu variant, so the tests becomes part of the qemu.
With this change, the ci-scripts/yocto.groovy script in
pelux-manifest also needs to be updated to remove the
automatic copying of test config to the local.conf file.

Signed-off-by: Tariq Ansari <tansari@luxoft.com>