pelux.conf: set DISTRO_VERSION by running git describe
In the manifests repository. Will make it so it is possible to know what
revision of pelux-manfiests was used when building an image.

Assign DISTRO_VERSION with := to avoid running git describe every time
DISTRO_VERSION is evaluated.

COREBASE points to the sources/poky directory of a PELUX checkout. Using
"../../.repo/manifests" means that we rely on the current folder structure
of a PELUX checkout to find the repository to run git describe in. Should
be OK. If it ever changes we probably need to modify other paths, this will
just be one more. A warning is printed by BitBake if the repository is not
found and a default string is used instead.

Since the Yocto version is used in the names of tags in pelux-manifests,
remove it from VERSION in os-release.bbappend by removing DISTRO_CODENAME.
Example of /usr/lib/os-release generated with this change:

ID="pelux"
NAME="PELUX"
VERSION="1.0-pyro-200-g17d4519 (Based on Yocto)"
VERSION_ID="1.0-pyro-200-g17d4519"
PRETTY_NAME="PELUX 1.0-pyro-200-g17d4519 (Based on Yocto)"

If a build is made on a tag, the number of commits following a tag ("-200")
and the short hash ("-g17d4519") will not be included. E.g. VERSION_ID will
be "<Major>.<Minor>-<Yocto code name>" for releases.

v1.0/pyro is currently the only tag we have on the master branch so the
version will be "1.0-pyro..." until that changes.

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>