Include mount parts of udev-extraconf in Poky
Will make it so pluggable mass storage devices are automatically mounted
under /run/media/ when inserted.

Since PELUX has two root partitions for swupdate, one that is actually
mounted at / and one that is not mounted at all, they need to be
blacklisted from the script that is run when a mass storage device is
detected. If it is not, the rootfs that is not / will be mounted under
/run/media/. mount.sh in udev-extraconf is extended with the functionality
to use UUID:s or labels to blacklist devices so that this can be reliable
done in PELUX (using the platform1 and platform2 labels).

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>