core-image-pelux: reduce number of IMAGE_INSTALL_append
Unnecessary to have one for each component. Also, most comments added on
the line above do not provide any extra information, so remove many of
them.

Create one "basic" group at top and one group for components that should
probably be included through e.g. dependencies or a feature. Not sure.
Perhaps udev-extraconf and swupdate should be in the last group as well?
Punt that for now and focus on limiting number of IMAGE_INSTALL_append.

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>