Rename /media partition to /data
Using /media is a bit unfortunate since it has historically been used to
mount removable devices in Linux. /run/media is the preferred location
now, and it is also what PELUX uses, but /media is still mentioned in old
standards (e.g. the FHS) and used in some Linux distributions (e.g.
Ubuntu).

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>