Cleanup /etc/fstab handling for Intel and ARP
By letting wic generate it once more and by using --use-uuid to generate
PARTUUID:s.

--use-uuid will cause "UUID=" (file system UUID) to be used in the
generated /etc/fstab compared to "PARTUUID=" in the hard coded /etc/fstab.
There is no way to force use of PARTUUID in wic when an fs UUID exists.
Could perhaps be argued that UUID is closer to the data and therefor better
to use. Example of generated /etc/fstab after this change:

/dev/root            /                    auto       defaults              1  1
UUID=F802-327C	/boot	vfat	defaults	0	0
UUID=e7247f57-63e9-4157-a1bf-6153b3dc60dd	swap	swap	defaults	0	0
UUID=2944d0a5-6b1c-49ac-b93e-cea99d272d12	/data	ext4	defaults	0	0

A static /etc/fstab file was added in ae64f26bdf2488ff93b7 to work around
the fact that /boot was not in /etc/fstab for one or both of the root
partitions (unclear). Have verified that /boot is included in /etc/fstab
on both root partitions after this change (the example above is extracted
from partitions of a *.wic ARP image).

a0827f57a2313b256e6f made it so "--no-fstab-update" is passed to wic. Want
/etc/fstab to be generated so remove it.

Hardcoded UUID:s were added in 52395463f867ae3c26e0 . Can be removed for
most partitions when --use-uuid is used since it causes ids to be
generated. Have to keep UUID:s for root partitions due to
"bootloader --configfile cfg" on in wks file. wic inserts generated
PARTUUID for root device when bootloader config is generated but there does
not seem to be a way to inject the generated PARTUUID in a pre-made file.

meta-arp-extras/ was added in b5219d7b55ac35547471 only to work around
the fact that /etc/fstab for Intel was not used for ARP. Remove it since
it is no longer needed.

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>