Revert "Cleanup /etc/fstab handling for Intel and ARP"
This reverts commit 1acf4025a35a630a572d31938ada529951d6c153.

Even though automatically generating fstab from the wks file
is a nice solution, but it does not work well with the swupdate
feature. The reason being that the swupdate swu file does not
contain the fstab file that is generated for wic image. And there
seems to be no way of updating the swu file to contain the fstab
generated by wic script. So the swu file uses the base fstab from
the base-files in poky layer. This results in the partitions mounted
incorrectly when the partition is updated using swupdate. So here we
revert to the solution that uses hard coded fstab files instead.