fstab: mount boot and data partition
RPI image was changed to wic format as well as
a new data partition was added to the image in
the following commits,

26079f1d6af92298494754f8d146cdcdc3edbf69
0d70eaf4fb638504ed9ed8bb5b09e4efae4fda60

fstab was not updated to mount the partitions
added to the wks file. This commit adds these
partitions to fstab.

WIC_CREATE_EXTRA_ARGS = "--no-fstab-update" is
also added to the local.conf for rpi. This stops
wic script to generated fstab file from wks file.
This is required to support swupdate swu image to
use the hardcoded fstab file instead.

Also the --uuid is removed from wks, as uuid is
not used to identify boot dev. Instead labels are
used to identify devices.