variant/smarcimx8m: mask connman append from meta-fsl-bsp-release
meta-fsl-bsp-release has bbappend for connman that supplies some
enhanced verison of connman script with an aim to fix few issues, e.g.
"issue that 2 Ethernet port board failed to mount nfs rootfs" that
are of no concern for us.

Mask this bbappend, so it doesn't conflict with bbappend in meta-pelux.

Signed-off-by: Oleksandr Kravchuk <oleksandr.kravchuk@pelagicore.com>