swupdate: disable systemd service
swupdate is currently configured to use one of the local development
workstation as a server which is not always available.

Moreover, we do not want swupdate to check for any updates on start, as
it is more of a proof of concept than a production ready solution, so
disable the corresponding systemd service by default.

It was once disabled in 73c3163ef5, but was later reenabled in
6f7e4f29e6.

Signed-off-by: Oleksandr Kravchuk <oleksandr.kravchuk@pelagicore.com>