neptune3-ui: fix setting brcm driver for rpi
meta-boot2 refactored systemd unit file for Neptune 3 UI, so now the
environment variables are stored in a separate file (bc3ff11632).

In be19aed09c we have added own environment file, but the path to the
file in which QT_WAYLAND_HARDWARE_INTEGRATION environment variable is
set was not changed. This commit fixes it.

Signed-off-by: Oleksandr Kravchuk <oleksandr.kravchuk@pelagicore.com>