layer.conf: use BBFILES_DYNAMIC to handle dynamic-layers
BBFILES_DYNAMIC is a native bitkbake feature that can be used to
activate content only when some other layers are present. The other
layers are identified by the collections that they define.

Use this feature instead of manually going through all the recipes
and appends and adding them only if the corresponding layer is
included in our configuration.

Also renamed layers to dynamic-layers as it is a common practice
to assign this name to the folder in the Yocto community.

For more info see 337bea6573 in poky or the Yocto Project Reference
Manual[1].

[1] https://yoctoproject.org/docs/latest/ref-manual/ref-manual.html#var-BBFILES_DYNAMIC

Signed-off-by: Oleksandr Kravchuk <oleksandr.kravchuk@pelagicore.com>