variant/arp: rename to arp-intel
Since the current ARP variant is virtually 99% identical to the Intel
one, we get rid of meta-arp and therefore "arp" machine type.

To be able to add additional packages to ARP iamges, we keep a separate
variant for it.

Because the current ARP variant is for Intel SMARC modules, we rename
it to arp-intel.

Signed-off-by: Oleksandr Kravchuk <oleksandr.kravchuk@pelagicore.com>