Move variant local.conf.sample content
Move the content in conf/variant/*/local.conf.sample to either
conf/variant/common/local.conf or classes/pelux.bbclass.

This is the first step in restructuring how variants are handled. The
end goal is to remove the conf/variant structure as it is way too
cumbersome to maintain.

Changing from "+=" to "_append" to align formatting.